Newer website content and plugins, added warnings
I've stopped using this, so I moved in some of the newer website changes
that I made a while ago and a couple more themes and plugins. I added
the PHP search in here as a reference. It is horribly coded, but may be
able to give some pointers to people. In addition, I added a readme file
and a script to convert from using this HTML generator to Jekyll.